Refactor the code a bit and created folders for better code visibility and readability
