+++ src/app/(tools)/svg-to-png/SVGRenderer.tsx
@@ -1,0 +1,24 @@
+import React from "react";
+
+interface SVGRendererProps {
+  svgContent: string;
+}
+
+const SVGRenderer: React.FC<SVGRendererProps> = ({ svgContent }) => {
+  const containerRef = React.useRef<HTMLDivElement>(null);
+
+  React.useEffect(() => {
+    if (containerRef.current) {
+      containerRef.current.innerHTML = svgContent;
+      const svgElement = containerRef.current.querySelector("svg");
+      if (svgElement) {
+        svgElement.setAttribute("width", "100%");
+        svgElement.setAttribute("height", "auto");
+      }
+    }
+  }, [svgContent]);
+
+  return <div ref={containerRef} />;
+};
+
+export default SVGRenderer;
