+++ src/app/(tools)/svg-to-png/SaveAsPngButton.tsx
@@ -1,0 +1,43 @@
+import React from "react";
+import useSvgConverter from "./hook/useSvgConverter";
+import type { Scale } from "@/app/utils/types";
+import { usePlausible } from "next-plausible";
+
+function SaveAsPngButton({
+  svgContent,
+  scale,
+  imageMetadata,
+}: {
+  svgContent: string;
+  scale: Scale;
+  imageMetadata: { width: number; height: number; name: string };
+}) {
+  const [canvasRef, setCanvasRef] = React.useState<HTMLCanvasElement | null>(
+    null,
+  );
+  const { convertToPng, canvasProps } = useSvgConverter({
+    canvas: canvasRef,
+    svgContent,
+    scale,
+    imageMetadata,
+  });
+
+  const plausible = usePlausible();
+
+  return (
+    <div>
+      <canvas ref={setCanvasRef} {...canvasProps} hidden />
+      <button
+        onClick={() => {
+          plausible("convert-svg-to-png");
+          void convertToPng();
+        }}
+        className="rounded-lg bg-green-700 px-4 py-2 text-sm font-semibold text-white shadow-md transition-colors duration-200 hover:bg-green-800 focus:outline-none focus:ring-2 focus:ring-green-400 focus:ring-opacity-75"
+      >
+        Save as PNG
+      </button>
+    </div>
+  );
+}
+
+export default SaveAsPngButton;
