+++ src/app/(tools)/svg-to-png/hook/useSvgConverter.tsx
@@ -1,0 +1,56 @@
+import  { useMemo } from "react";
+import { scaleSvg } from "@/app/utils/helper";
+import type { Scale } from "@/app/utils/types";
+
+function useSvgConverter(props: {
+  canvas: HTMLCanvasElement | null;
+  svgContent: string;
+  scale: Scale;
+  fileName?: string;
+  imageMetadata: { width: number; height: number; name: string };
+}) {
+  const { width, height, scaledSvg } = useMemo(() => {
+    const scaledSvg = scaleSvg(props.svgContent, props.scale);
+
+    return {
+      width: props.imageMetadata.width * props.scale,
+      height: props.imageMetadata.height * props.scale,
+      scaledSvg,
+    };
+  }, [props.svgContent, props.scale, props.imageMetadata]);
+
+  const convertToPng = async () => {
+    const ctx = props.canvas?.getContext("2d");
+    if (!ctx) throw new Error("Failed to get canvas context");
+
+    // Trigger a "save image" of the resulting canvas content
+    const saveImage = () => {
+      if (props.canvas) {
+        const dataURL = props.canvas.toDataURL("image/png");
+        const link = document.createElement("a");
+        link.href = dataURL;
+        const svgFileName = props.imageMetadata.name ?? "svg_converted";
+
+        // Remove the .svg extension
+        link.download = `${svgFileName.replace(".svg", "")}-${props.scale}x.png`;
+        link.click();
+      }
+    };
+
+    const img = new Image();
+    // Call saveImage after the image has been drawn
+    img.onload = () => {
+      ctx.drawImage(img, 0, 0);
+      saveImage();
+    };
+
+    img.src = `data:image/svg+xml;charset=utf-8,${encodeURIComponent(scaledSvg)}`;
+  };
+
+  return {
+    convertToPng,
+    canvasProps: { width: width, height: height },
+  };
+}
+
+export default useSvgConverter;
