+++ src/app/utils/helper.tsx
@@ -1,0 +1,17 @@
+import type { Scale } from "./types";
+
+export function scaleSvg(svgContent: string, scale: Scale) {
+  const parser = new DOMParser();
+  const svgDoc = parser.parseFromString(svgContent, "image/svg+xml");
+  const svgElement = svgDoc.documentElement;
+  const width = parseInt(svgElement.getAttribute("width") ?? "300");
+  const height = parseInt(svgElement.getAttribute("height") ?? "150");
+
+  const scaledWidth = width * scale;
+  const scaledHeight = height * scale;
+
+  svgElement.setAttribute("width", scaledWidth.toString());
+  svgElement.setAttribute("height", scaledHeight.toString());
+
+  return new XMLSerializer().serializeToString(svgDoc);
+}
