+++ src/app/utils/types.tsx
@@ -1,0 +1,1 @@
+export type Scale = 1 | 2 | 4 | 8 | 16 | 32 | 64;
